update Trip functionality

+++ src/main/java/com/tsystems/railway/DTO/StationDTO.java
@@ -3,16 +3,16 @@ package com.tsystems.railway.DTO;
 import com.tsystems.railway.model.Station;
 
 public class StationDTO {
-    private long id;
+    private int id;
 
     private String name;
 
-    public StationDTO(long id, String name) {
+    public StationDTO(int id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
